refactor: decompose case-classifier (#208)
- Extract word validators to sentence-case/word-validators.js
- Extract bold text logic to sentence-case/bold-text-classifier.js
- Reduce case-classifier.js from 1063 to 371 LOC per ADR-004
